Cleanup: reduce helpers/math (#3511)
